reorg API
