+++ server/api/api.go
@@ -85,13 +85,10 @@ func (a *API) RegisterRoutes(r *mux.Router) {
 	apiv2.Use(a.panicHandler)
 	apiv2.Use(a.requireCSRFToken)
 
-	// personal-server specific routes. These are not needed in plugin mode.
-	if !a.isPlugin {
-		apiv2.HandleFunc("/login", a.handleLogin).Methods("POST")
-		apiv2.HandleFunc("/logout", a.sessionRequired(a.handleLogout)).Methods("POST")
-		apiv2.HandleFunc("/register", a.handleRegister).Methods("POST")
-		apiv2.HandleFunc("/teams/{teamID}/regenerate_signup_token", a.sessionRequired(a.handlePostTeamRegenerateSignupToken)).Methods("POST")
-	}
+	a.registerUsersRoutes(apiv2)
+	a.registerAuthRoutes(apiv2)
+	a.registerMembersRoutes(apiv2)
+	a.registerCategoriesRoutes(apiv2)
 
 	// Board APIs
 	apiv2.HandleFunc("/teams/{teamID}/channels", a.sessionRequired(a.handleSearchMyChannels)).Methods("GET")
@@ -116,14 +113,6 @@ func (a *API) RegisterRoutes(r *mux.Router) {
 	apiv2.HandleFunc("/boards/{boardID}/blocks/{blockID}/duplicate", a.sessionRequired(a.handleDuplicateBlock)).Methods("POST")
 	apiv2.HandleFunc("/boards/{boardID}/metadata", a.sessionRequired(a.handleGetBoardMetadata)).Methods("GET")
 
-	// Member APIs
-	apiv2.HandleFunc("/boards/{boardID}/members", a.sessionRequired(a.handleGetMembersForBoard)).Methods("GET")
-	apiv2.HandleFunc("/boards/{boardID}/members", a.sessionRequired(a.handleAddMember)).Methods("POST")
-	apiv2.HandleFunc("/boards/{boardID}/members/{userID}", a.sessionRequired(a.handleUpdateMember)).Methods("PUT")
-	apiv2.HandleFunc("/boards/{boardID}/members/{userID}", a.sessionRequired(a.handleDeleteMember)).Methods("DELETE")
-	apiv2.HandleFunc("/boards/{boardID}/join", a.sessionRequired(a.handleJoinBoard)).Methods("POST")
-	apiv2.HandleFunc("/boards/{boardID}/leave", a.sessionRequired(a.handleLeaveBoard)).Methods("POST")
-
 	// Sharing APIs
 	apiv2.HandleFunc("/boards/{boardID}/sharing", a.sessionRequired(a.handlePostSharing)).Methods("POST")
 	apiv2.HandleFunc("/boards/{boardID}/sharing", a.sessionRequired(a.handleGetSharing)).Methods("GET")
@@ -135,14 +124,6 @@ func (a *API) RegisterRoutes(r *mux.Router) {
 	apiv2.HandleFunc("/teams/{teamID}/archive/export", a.sessionRequired(a.handleArchiveExportTeam)).Methods("GET")
 	apiv2.HandleFunc("/teams/{teamID}/{boardID}/files", a.sessionRequired(a.handleUploadFile)).Methods("POST")
 
-	// User APIs
-	apiv2.HandleFunc("/users", a.sessionRequired(a.handleGetUsersList)).Methods("POST")
-	apiv2.HandleFunc("/users/me", a.sessionRequired(a.handleGetMe)).Methods("GET")
-	apiv2.HandleFunc("/users/me/memberships", a.sessionRequired(a.handleGetMyMemberships)).Methods("GET")
-	apiv2.HandleFunc("/users/{userID}", a.sessionRequired(a.handleGetUser)).Methods("GET")
-	apiv2.HandleFunc("/users/{userID}/changepassword", a.sessionRequired(a.handleChangePassword)).Methods("POST")
-	apiv2.HandleFunc("/users/{userID}/config", a.sessionRequired(a.handleUpdateUserConfig)).Methods(http.MethodPut)
-
 	// BoardsAndBlocks APIs
 	apiv2.HandleFunc("/boards-and-blocks", a.sessionRequired(a.handleCreateBoardsAndBlocks)).Methods("POST")
 	apiv2.HandleFunc("/boards-and-blocks", a.sessionRequired(a.handlePatchBoardsAndBlocks)).Methods("PATCH")
@@ -150,15 +131,6 @@ func (a *API) RegisterRoutes(r *mux.Router) {
 
 	// Auth APIs
 	apiv2.HandleFunc("/clientConfig", a.getClientConfig).Methods("GET")
-
-	// Category APIs
-	apiv2.HandleFunc("/teams/{teamID}/categories", a.sessionRequired(a.handleCreateCategory)).Methods(http.MethodPost)
-	apiv2.HandleFunc("/teams/{teamID}/categories/{categoryID}", a.sessionRequired(a.handleUpdateCategory)).Methods(http.MethodPut)
-	apiv2.HandleFunc("/teams/{teamID}/categories/{categoryID}", a.sessionRequired(a.handleDeleteCategory)).Methods(http.MethodDelete)
-
-	// Category Block APIs
-	apiv2.HandleFunc("/teams/{teamID}/categories", a.sessionRequired(a.handleGetUserCategoryBoards)).Methods(http.MethodGet)
-	apiv2.HandleFunc("/teams/{teamID}/categories/{categoryID}/boards/{boardID}", a.sessionRequired(a.handleUpdateCategoryBoard)).Methods(http.MethodPost)
 
 	// Get Files API
 	apiv2.HandleFunc("/files/teams/{teamID}/{boardID}/{filename}", a.attachSession(a.handleServeFile, false)).Methods("GET")
@@ -424,384 +396,6 @@ func (a *API) handleGetBlocks(w http.ResponseWriter, r *http.Request) {
 	auditRec.Success()
 }
 
-func (a *API) handleCreateCategory(w http.ResponseWriter, r *http.Request) {
-	// swagger:operation POST /teams/{teamID}/categories createCategory
-	//
-	// Create a category for boards
-	//
-	// ---
-	// produces:
-	// - application/json
-	// parameters:
-	// - name: teamID
-	//   in: path
-	//   description: Team ID
-	//   required: true
-	//   type: string
-	// - name: Body
-	//   in: body
-	//   description: category to create
-	//   required: true
-	//   schema:
-	//     "$ref": "#/definitions/Category"
-	// security:
-	// - BearerAuth: []
-	// responses:
-	//   '200':
-	//     description: success
-	//     schema:
-	//       "$ref": "#/definitions/Category"
-	//   default:
-	//     description: internal error
-	//     schema:
-	//       "$ref": "#/definitions/ErrorResponse"
-
-	requestBody, err := ioutil.ReadAll(r.Body)
-	if err != nil {
-		a.errorResponse(w, r.URL.Path, http.StatusInternalServerError, "", err)
-		return
-	}
-
-	var category model.Category
-
-	err = json.Unmarshal(requestBody, &category)
-	if err != nil {
-		a.errorResponse(w, r.URL.Path, http.StatusInternalServerError, "", err)
-		return
-	}
-
-	auditRec := a.makeAuditRecord(r, "createCategory", audit.Fail)
-	defer a.audit.LogRecord(audit.LevelModify, auditRec)
-
-	ctx := r.Context()
-	session := ctx.Value(sessionContextKey).(*model.Session)
-
-	// user can only create category for themselves
-	if category.UserID != session.UserID {
-		a.errorResponse(
-			w,
-			r.URL.Path,
-			http.StatusBadRequest,
-			fmt.Sprintf("userID %s and category userID %s mismatch", session.UserID, category.UserID),
-			nil,
-		)
-		return
-	}
-
-	vars := mux.Vars(r)
-	teamID := vars["teamID"]
-
-	if category.TeamID != teamID {
-		a.errorResponse(
-			w,
-			r.URL.Path,
-			http.StatusBadRequest,
-			"teamID mismatch",
-			nil,
-		)
-		return
-	}
-
-	createdCategory, err := a.app.CreateCategory(&category)
-	if err != nil {
-		a.errorResponse(w, r.URL.Path, http.StatusInternalServerError, "", err)
-		return
-	}
-
-	data, err := json.Marshal(createdCategory)
-	if err != nil {
-		a.errorResponse(w, r.URL.Path, http.StatusInternalServerError, "", err)
-		return
-	}
-
-	jsonBytesResponse(w, http.StatusOK, data)
-	auditRec.AddMeta("categoryID", createdCategory.ID)
-	auditRec.Success()
-}
-
-func (a *API) handleUpdateCategory(w http.ResponseWriter, r *http.Request) {
-	// swagger:operation PUT /teams/{teamID}/categories/{categoryID} updateCategory
-	//
-	// Create a category for boards
-	//
-	// ---
-	// produces:
-	// - application/json
-	// parameters:
-	// - name: teamID
-	//   in: path
-	//   description: Team ID
-	//   required: true
-	//   type: string
-	// - name: categoryID
-	//   in: path
-	//   description: Category ID
-	//   required: true
-	//   type: string
-	// - name: Body
-	//   in: body
-	//   description: category to update
-	//   required: true
-	//   schema:
-	//     "$ref": "#/definitions/Category"
-	// security:
-	// - BearerAuth: []
-	// responses:
-	//   '200':
-	//     description: success
-	//     schema:
-	//       "$ref": "#/definitions/Category"
-	//   default:
-	//     description: internal error
-	//     schema:
-	//       "$ref": "#/definitions/ErrorResponse"
-
-	vars := mux.Vars(r)
-	categoryID := vars["categoryID"]
-
-	requestBody, err := ioutil.ReadAll(r.Body)
-	if err != nil {
-		a.errorResponse(w, r.URL.Path, http.StatusInternalServerError, "", err)
-		return
-	}
-
-	var category model.Category
-	err = json.Unmarshal(requestBody, &category)
-	if err != nil {
-		a.errorResponse(w, r.URL.Path, http.StatusInternalServerError, "", err)
-		return
-	}
-
-	auditRec := a.makeAuditRecord(r, "updateCategory", audit.Fail)
-	defer a.audit.LogRecord(audit.LevelModify, auditRec)
-
-	if categoryID != category.ID {
-		a.errorResponse(w, r.URL.Path, http.StatusBadRequest, "categoryID mismatch in patch and body", nil)
-		return
-	}
-
-	ctx := r.Context()
-	session := ctx.Value(sessionContextKey).(*model.Session)
-
-	// user can only update category for themselves
-	if category.UserID != session.UserID {
-		a.errorResponse(w, r.URL.Path, http.StatusBadRequest, "user ID mismatch in session and category", nil)
-		return
-	}
-
-	teamID := vars["teamID"]
-	if category.TeamID != teamID {
-		a.errorResponse(
-			w,
-			r.URL.Path,
-			http.StatusBadRequest,
-			"teamID mismatch",
-			nil,
-		)
-		return
-	}
-
-	updatedCategory, err := a.app.UpdateCategory(&category)
-	if err != nil {
-		switch {
-		case errors.Is(err, app.ErrorCategoryDeleted):
-			a.errorResponse(w, r.URL.Path, http.StatusNotFound, "", err)
-		case errors.Is(err, app.ErrorCategoryPermissionDenied):
-			// TODO: The permissions should be handled as much as possible at
-			// the API level, this needs to be changed
-			a.errorResponse(w, r.URL.Path, http.StatusForbidden, "", err)
-		default:
-			a.errorResponse(w, r.URL.Path, http.StatusInternalServerError, "", err)
-		}
-		return
-	}
-
-	data, err := json.Marshal(updatedCategory)
-	if err != nil {
-		a.errorResponse(w, r.URL.Path, http.StatusInternalServerError, "", err)
-		return
-	}
-
-	jsonBytesResponse(w, http.StatusOK, data)
-	auditRec.Success()
-}
-
-func (a *API) handleDeleteCategory(w http.ResponseWriter, r *http.Request) {
-	// swagger:operation DELETE /teams/{teamID}/categories/{categoryID} deleteCategory
-	//
-	// Delete a category
-	//
-	// ---
-	// produces:
-	// - application/json
-	// parameters:
-	// - name: teamID
-	//   in: path
-	//   description: Team ID
-	//   required: true
-	//   type: string
-	// - name: categoryID
-	//   in: path
-	//   description: Category ID
-	//   required: true
-	//   type: string
-	// security:
-	// - BearerAuth: []
-	// responses:
-	//   '200':
-	//     description: success
-	//   default:
-	//     description: internal error
-	//     schema:
-	//       "$ref": "#/definitions/ErrorResponse"
-
-	ctx := r.Context()
-	session := ctx.Value(sessionContextKey).(*model.Session)
-	vars := mux.Vars(r)
-
-	userID := session.UserID
-	teamID := vars["teamID"]
-	categoryID := vars["categoryID"]
-
-	auditRec := a.makeAuditRecord(r, "deleteCategory", audit.Fail)
-	defer a.audit.LogRecord(audit.LevelModify, auditRec)
-
-	deletedCategory, err := a.app.DeleteCategory(categoryID, userID, teamID)
-	if err != nil {
-		switch {
-		case errors.Is(err, app.ErrorInvalidCategory):
-			a.errorResponse(w, r.URL.Path, http.StatusBadRequest, "", err)
-		case errors.Is(err, app.ErrorCategoryPermissionDenied):
-			// TODO: The permissions should be handled as much as possible at
-			// the API level, this needs to be changed
-			a.errorResponse(w, r.URL.Path, http.StatusForbidden, "", err)
-		default:
-			a.errorResponse(w, r.URL.Path, http.StatusInternalServerError, "", err)
-		}
-		return
-	}
-
-	data, err := json.Marshal(deletedCategory)
-	if err != nil {
-		a.errorResponse(w, r.URL.Path, http.StatusInternalServerError, "", err)
-		return
-	}
-
-	jsonBytesResponse(w, http.StatusOK, data)
-	auditRec.Success()
-}
-
-func (a *API) handleGetUserCategoryBoards(w http.ResponseWriter, r *http.Request) {
-	// swagger:operation GET /teams/{teamID}/categories getUserCategoryBoards
-	//
-	// Gets the user's board categories
-	//
-	// ---
-	// produces:
-	// - application/json
-	// parameters:
-	// - name: teamID
-	//   in: path
-	//   description: Team ID
-	//   required: true
-	//   type: string
-	// security:
-	// - BearerAuth: []
-	// responses:
-	//   '200':
-	//     description: success
-	//     schema:
-	//       items:
-	//         "$ref": "#/definitions/CategoryBoards"
-	//       type: array
-	//   default:
-	//     description: internal error
-	//     schema:
-	//       "$ref": "#/definitions/ErrorResponse"
-
-	ctx := r.Context()
-	session := ctx.Value(sessionContextKey).(*model.Session)
-	userID := session.UserID
-
-	vars := mux.Vars(r)
-	teamID := vars["teamID"]
-
-	auditRec := a.makeAuditRecord(r, "getUserCategoryBoards", audit.Fail)
-	defer a.audit.LogRecord(audit.LevelModify, auditRec)
-
-	categoryBlocks, err := a.app.GetUserCategoryBoards(userID, teamID)
-	if err != nil {
-		a.errorResponse(w, r.URL.Path, http.StatusInternalServerError, "", err)
-		return
-	}
-
-	data, err := json.Marshal(categoryBlocks)
-	if err != nil {
-		a.errorResponse(w, r.URL.Path, http.StatusInternalServerError, "", err)
-		return
-	}
-
-	jsonBytesResponse(w, http.StatusOK, data)
-	auditRec.Success()
-}
-
-func (a *API) handleUpdateCategoryBoard(w http.ResponseWriter, r *http.Request) {
-	// swagger:operation POST /teams/{teamID}/categories/{categoryID}/boards/{boardID} updateCategoryBoard
-	//
-	// Set the category of a board
-	//
-	// ---
-	// produces:
-	// - application/json
-	// parameters:
-	// - name: teamID
-	//   in: path
-	//   description: Team ID
-	//   required: true
-	//   type: string
-	// - name: categoryID
-	//   in: path
-	//   description: Category ID
-	//   required: true
-	//   type: string
-	// - name: boardID
-	//   in: path
-	//   description: Board ID
-	//   required: true
-	//   type: string
-	// security:
-	// - BearerAuth: []
-	// responses:
-	//   '200':
-	//     description: success
-	//   default:
-	//     description: internal error
-	//     schema:
-	//       "$ref": "#/definitions/ErrorResponse"
-
-	auditRec := a.makeAuditRecord(r, "updateCategoryBoard", audit.Fail)
-	defer a.audit.LogRecord(audit.LevelModify, auditRec)
-
-	vars := mux.Vars(r)
-	categoryID := vars["categoryID"]
-	boardID := vars["boardID"]
-	teamID := vars["teamID"]
-
-	ctx := r.Context()
-	session := ctx.Value(sessionContextKey).(*model.Session)
-	userID := session.UserID
-
-	// TODO: Check the category and the team matches
-	err := a.app.AddUpdateUserCategoryBoard(teamID, userID, categoryID, boardID)
-	if err != nil {
-		a.errorResponse(w, r.URL.Path, http.StatusInternalServerError, "", err)
-		return
-	}
-
-	jsonBytesResponse(w, http.StatusOK, []byte("ok"))
-	auditRec.Success()
-}
-
 func (a *API) handlePostBlocks(w http.ResponseWriter, r *http.Request) {
 	// swagger:operation POST /boards/{boardID}/blocks updateBlocks
 	//
@@ -944,288 +538,6 @@ func (a *API) handlePostBlocks(w http.ResponseWriter, r *http.Request) {
 	jsonBytesResponse(w, http.StatusOK, json)
 
 	auditRec.AddMeta("blockCount", len(blocks))
-	auditRec.Success()
-}
-
-func (a *API) handleUpdateUserConfig(w http.ResponseWriter, r *http.Request) {
-	// swagger:operation PATCH /users/{userID}/config updateUserConfig
-	//
-	// Updates user config
-	//
-	// ---
-	// produces:
-	// - application/json
-	// parameters:
-	// - name: userID
-	//   in: path
-	//   description: User ID
-	//   required: true
-	//   type: string
-	// - name: Body
-	//   in: body
-	//   description: User config patch to apply
-	//   required: true
-	//   schema:
-	//     "$ref": "#/definitions/UserPropPatch"
-	// security:
-	// - BearerAuth: []
-	// responses:
-	//   '200':
-	//     description: success
-	//   default:
-	//     description: internal error
-	//     schema:
-	//       "$ref": "#/definitions/ErrorResponse"
-
-	requestBody, err := ioutil.ReadAll(r.Body)
-	if err != nil {
-		a.errorResponse(w, r.URL.Path, http.StatusInternalServerError, "", err)
-		return
-	}
-
-	var patch *model.UserPropPatch
-	err = json.Unmarshal(requestBody, &patch)
-	if err != nil {
-		a.errorResponse(w, r.URL.Path, http.StatusInternalServerError, "", err)
-		return
-	}
-
-	vars := mux.Vars(r)
-	userID := vars["userID"]
-
-	ctx := r.Context()
-	session := ctx.Value(sessionContextKey).(*model.Session)
-
-	auditRec := a.makeAuditRecord(r, "updateUserConfig", audit.Fail)
-	defer a.audit.LogRecord(audit.LevelModify, auditRec)
-
-	// a user can update only own config
-	if userID != session.UserID {
-		a.errorResponse(w, r.URL.Path, http.StatusForbidden, "", nil)
-		return
-	}
-
-	updatedConfig, err := a.app.UpdateUserConfig(userID, *patch)
-	if err != nil {
-		a.errorResponse(w, r.URL.Path, http.StatusInternalServerError, "", err)
-		return
-	}
-
-	data, err := json.Marshal(updatedConfig)
-	if err != nil {
-		a.errorResponse(w, r.URL.Path, http.StatusInternalServerError, "", err)
-		return
-	}
-
-	jsonBytesResponse(w, http.StatusOK, data)
-	auditRec.Success()
-}
-
-func (a *API) handleGetUser(w http.ResponseWriter, r *http.Request) {
-	// swagger:operation GET /users/{userID} getUser
-	//
-	// Returns a user
-	//
-	// ---
-	// produces:
-	// - application/json
-	// parameters:
-	// - name: userID
-	//   in: path
-	//   description: User ID
-	//   required: true
-	//   type: string
-	// security:
-	// - BearerAuth: []
-	// responses:
-	//   '200':
-	//     description: success
-	//     schema:
-	//       "$ref": "#/definitions/User"
-	//   default:
-	//     description: internal error
-	//     schema:
-	//       "$ref": "#/definitions/ErrorResponse"
-
-	vars := mux.Vars(r)
-	userID := vars["userID"]
-
-	auditRec := a.makeAuditRecord(r, "postBlocks", audit.Fail)
-	defer a.audit.LogRecord(audit.LevelRead, auditRec)
-	auditRec.AddMeta("userID", userID)
-
-	user, err := a.app.GetUser(userID)
-	if err != nil {
-		a.errorResponse(w, r.URL.Path, http.StatusInternalServerError, "", err)
-		return
-	}
-
-	userData, err := json.Marshal(user)
-	if err != nil {
-		a.errorResponse(w, r.URL.Path, http.StatusInternalServerError, "", err)
-		return
-	}
-
-	jsonBytesResponse(w, http.StatusOK, userData)
-	auditRec.Success()
-}
-
-func (a *API) handleGetUsersList(w http.ResponseWriter, r *http.Request) {
-	// swagger:operation POST /users getUser
-	//
-	// Returns a user[]
-	//
-	// ---
-	// produces:
-	// - application/json
-	// parameters:
-	// - name: userID
-	//   in: path
-	//   description: User ID
-	//   required: true
-	//   type: string
-	// security:
-	// - BearerAuth: []
-	// responses:
-	//   '200':
-	//     description: success
-	//     schema:
-	//       "$ref": "#/definitions/User"
-	//   default:
-	//     description: internal error
-	//     schema:
-	//       "$ref": "#/definitions/ErrorResponse"
-
-	requestBody, err := ioutil.ReadAll(r.Body)
-	if err != nil {
-		a.errorResponse(w, r.URL.Path, http.StatusInternalServerError, "", err)
-		return
-	}
-
-	var userIDs []string
-	if err = json.Unmarshal(requestBody, &userIDs); err != nil {
-		a.errorResponse(w, r.URL.Path, http.StatusInternalServerError, "", err)
-		return
-	}
-
-	auditRec := a.makeAuditRecord(r, "getUsersList", audit.Fail)
-	defer a.audit.LogRecord(audit.LevelAuth, auditRec)
-
-	users, err := a.app.GetUsersList(userIDs)
-	if err != nil {
-		a.errorResponse(w, r.URL.Path, http.StatusBadRequest, err.Error(), err)
-		return
-	}
-
-	usersList, err := json.Marshal(users)
-	if err != nil {
-		a.errorResponse(w, r.URL.Path, http.StatusBadRequest, "", err)
-		return
-	}
-
-	jsonStringResponse(w, http.StatusOK, string(usersList))
-	auditRec.Success()
-}
-
-func (a *API) handleGetMe(w http.ResponseWriter, r *http.Request) {
-	// swagger:operation GET /users/me getMe
-	//
-	// Returns the currently logged-in user
-	//
-	// ---
-	// produces:
-	// - application/json
-	// security:
-	// - BearerAuth: []
-	// responses:
-	//   '200':
-	//     description: success
-	//     schema:
-	//       "$ref": "#/definitions/User"
-	//   default:
-	//     description: internal error
-	//     schema:
-	//       "$ref": "#/definitions/ErrorResponse"
-
-	userID := getUserID(r)
-
-	var user *model.User
-	var err error
-
-	auditRec := a.makeAuditRecord(r, "getMe", audit.Fail)
-	defer a.audit.LogRecord(audit.LevelRead, auditRec)
-
-	if userID == model.SingleUser {
-		ws, _ := a.app.GetRootTeam()
-		now := utils.GetMillis()
-		user = &model.User{
-			ID:       model.SingleUser,
-			Username: model.SingleUser,
-			Email:    model.SingleUser,
-			CreateAt: ws.UpdateAt,
-			UpdateAt: now,
-		}
-	} else {
-		user, err = a.app.GetUser(userID)
-		if err != nil {
-			a.errorResponse(w, r.URL.Path, http.StatusInternalServerError, "", err)
-			return
-		}
-	}
-
-	userData, err := json.Marshal(user)
-	if err != nil {
-		a.errorResponse(w, r.URL.Path, http.StatusInternalServerError, "", err)
-		return
-	}
-	jsonBytesResponse(w, http.StatusOK, userData)
-
-	auditRec.AddMeta("userID", user.ID)
-	auditRec.Success()
-}
-
-func (a *API) handleGetMyMemberships(w http.ResponseWriter, r *http.Request) {
-	// swagger:operation GET /users/me/memberships getMyMemberships
-	//
-	// Returns the currently users board memberships
-	//
-	// ---
-	// produces:
-	// - application/json
-	// security:
-	// - BearerAuth: []
-	// responses:
-	//   '200':
-	//     description: success
-	//     schema:
-	//       type: array
-	//       items:
-	//         "$ref": "#/definitions/BoardMember"
-	//   default:
-	//     description: internal error
-	//     schema:
-	//       "$ref": "#/definitions/ErrorResponse"
-
-	userID := getUserID(r)
-
-	auditRec := a.makeAuditRecord(r, "getMyBoardMemberships", audit.Fail)
-	auditRec.AddMeta("userID", userID)
-	defer a.audit.LogRecord(audit.LevelRead, auditRec)
-
-	members, err := a.app.GetMembersForUser(userID)
-	if err != nil {
-		a.errorResponse(w, r.URL.Path, http.StatusInternalServerError, "", err)
-		return
-	}
-
-	membersData, err := json.Marshal(members)
-	if err != nil {
-		a.errorResponse(w, r.URL.Path, http.StatusInternalServerError, "", err)
-		return
-	}
-
-	jsonBytesResponse(w, http.StatusOK, membersData)
-
 	auditRec.Success()
 }
 
@@ -3889,517 +3201,6 @@ func (a *API) handleSearchAllBoards(w http.ResponseWriter, r *http.Request) {
 	jsonBytesResponse(w, http.StatusOK, data)
 
 	auditRec.AddMeta("boardsCount", len(boards))
-	auditRec.Success()
-}
-
-func (a *API) handleGetMembersForBoard(w http.ResponseWriter, r *http.Request) {
-	// swagger:operation GET /boards/{boardID}/members getMembersForBoard
-	//
-	// Returns the members of the board
-	//
-	// ---
-	// produces:
-	// - application/json
-	// parameters:
-	// - name: boardID
-	//   in: path
-	//   description: Board ID
-	//   required: true
-	//   type: string
-	// security:
-	// - BearerAuth: []
-	// responses:
-	//   '200':
-	//     description: success
-	//     schema:
-	//       type: array
-	//       items:
-	//         "$ref": "#/definitions/BoardMember"
-	//   default:
-	//     description: internal error
-	//     schema:
-	//       "$ref": "#/definitions/ErrorResponse"
-
-	boardID := mux.Vars(r)["boardID"]
-	userID := getUserID(r)
-
-	if !a.permissions.HasPermissionToBoard(userID, boardID, model.PermissionViewBoard) {
-		a.errorResponse(w, r.URL.Path, http.StatusForbidden, "", PermissionError{"access denied to board members"})
-		return
-	}
-
-	auditRec := a.makeAuditRecord(r, "getMembersForBoard", audit.Fail)
-	defer a.audit.LogRecord(audit.LevelModify, auditRec)
-	auditRec.AddMeta("boardID", boardID)
-
-	members, err := a.app.GetMembersForBoard(boardID)
-	if err != nil {
-		a.errorResponse(w, r.URL.Path, http.StatusInternalServerError, "", err)
-		return
-	}
-
-	a.logger.Debug("GetMembersForBoard",
-		mlog.String("boardID", boardID),
-		mlog.Int("membersCount", len(members)),
-	)
-
-	data, err := json.Marshal(members)
-	if err != nil {
-		a.errorResponse(w, r.URL.Path, http.StatusInternalServerError, "", err)
-		return
-	}
-
-	// response
-	jsonBytesResponse(w, http.StatusOK, data)
-
-	auditRec.Success()
-}
-
-func (a *API) handleAddMember(w http.ResponseWriter, r *http.Request) {
-	// swagger:operation POST /boards/{boardID}/members addMember
-	//
-	// Adds a new member to a board
-	//
-	// ---
-	// produces:
-	// - application/json
-	// parameters:
-	// - name: boardID
-	//   in: path
-	//   description: Board ID
-	//   required: true
-	//   type: string
-	// - name: Body
-	//   in: body
-	//   description: membership to replace the current one with
-	//   required: true
-	//   schema:
-	//     "$ref": "#/definitions/BoardMember"
-	// security:
-	// - BearerAuth: []
-	// responses:
-	//   '200':
-	//     description: success
-	//     schema:
-	//       $ref: '#/definitions/BoardMember'
-	//   '404':
-	//     description: board not found
-	//   default:
-	//     description: internal error
-	//     schema:
-	//       "$ref": "#/definitions/ErrorResponse"
-
-	boardID := mux.Vars(r)["boardID"]
-	userID := getUserID(r)
-
-	board, err := a.app.GetBoard(boardID)
-	if err != nil {
-		a.errorResponse(w, r.URL.Path, http.StatusInternalServerError, "", err)
-		return
-	}
-	if board == nil {
-		a.errorResponse(w, r.URL.Path, http.StatusNotFound, "", nil)
-		return
-	}
-
-	if !a.permissions.HasPermissionToBoard(userID, boardID, model.PermissionManageBoardRoles) {
-		a.errorResponse(w, r.URL.Path, http.StatusForbidden, "", PermissionError{"access denied to modify board members"})
-		return
-	}
-
-	requestBody, err := ioutil.ReadAll(r.Body)
-	if err != nil {
-		a.errorResponse(w, r.URL.Path, http.StatusInternalServerError, "", err)
-		return
-	}
-
-	var reqBoardMember *model.BoardMember
-	if err = json.Unmarshal(requestBody, &reqBoardMember); err != nil {
-		a.errorResponse(w, r.URL.Path, http.StatusBadRequest, "", err)
-		return
-	}
-
-	if reqBoardMember.UserID == "" {
-		a.errorResponse(w, r.URL.Path, http.StatusBadRequest, "", err)
-		return
-	}
-
-	// currently all memberships are created as editors by default
-	newBoardMember := &model.BoardMember{
-		UserID:       reqBoardMember.UserID,
-		BoardID:      boardID,
-		SchemeEditor: true,
-	}
-
-	auditRec := a.makeAuditRecord(r, "addMember", audit.Fail)
-	defer a.audit.LogRecord(audit.LevelModify, auditRec)
-	auditRec.AddMeta("boardID", boardID)
-	auditRec.AddMeta("addedUserID", reqBoardMember.UserID)
-
-	member, err := a.app.AddMemberToBoard(newBoardMember)
-	if err != nil {
-		a.errorResponse(w, r.URL.Path, http.StatusInternalServerError, "", err)
-		return
-	}
-
-	a.logger.Debug("AddMember",
-		mlog.String("boardID", board.ID),
-		mlog.String("addedUserID", reqBoardMember.UserID),
-	)
-
-	data, err := json.Marshal(member)
-	if err != nil {
-		a.errorResponse(w, r.URL.Path, http.StatusInternalServerError, "", err)
-		return
-	}
-
-	// response
-	jsonBytesResponse(w, http.StatusOK, data)
-
-	auditRec.Success()
-}
-
-func (a *API) handleJoinBoard(w http.ResponseWriter, r *http.Request) {
-	// swagger:operation POST /boards/{boardID}/join joinBoard
-	//
-	// Become a member of a board
-	//
-	// ---
-	// produces:
-	// - application/json
-	// parameters:
-	// - name: boardID
-	//   in: path
-	//   description: Board ID
-	//   required: true
-	//   type: string
-	// security:
-	// - BearerAuth: []
-	// responses:
-	//   '200':
-	//     description: success
-	//     schema:
-	//       $ref: '#/definitions/BoardMember'
-	//   '404':
-	//     description: board not found
-	//   '403':
-	//     description: access denied
-	//   default:
-	//     description: internal error
-	//     schema:
-	//       "$ref": "#/definitions/ErrorResponse"
-
-	userID := getUserID(r)
-	if userID == "" {
-		a.errorResponse(w, r.URL.Path, http.StatusBadRequest, "", nil)
-		return
-	}
-
-	boardID := mux.Vars(r)["boardID"]
-	board, err := a.app.GetBoard(boardID)
-	if err != nil {
-		a.errorResponse(w, r.URL.Path, http.StatusInternalServerError, "", err)
-		return
-	}
-	if board == nil {
-		a.errorResponse(w, r.URL.Path, http.StatusNotFound, "", nil)
-		return
-	}
-	if board.Type != model.BoardTypeOpen {
-		a.errorResponse(w, r.URL.Path, http.StatusForbidden, "", nil)
-		return
-	}
-
-	if !a.permissions.HasPermissionToTeam(userID, board.TeamID, model.PermissionViewTeam) {
-		a.errorResponse(w, r.URL.Path, http.StatusForbidden, "", nil)
-		return
-	}
-
-	newBoardMember := &model.BoardMember{
-		UserID:          userID,
-		BoardID:         boardID,
-		SchemeAdmin:     board.MinimumRole == model.BoardRoleAdmin,
-		SchemeEditor:    board.MinimumRole == model.BoardRoleNone || board.MinimumRole == model.BoardRoleEditor,
-		SchemeCommenter: board.MinimumRole == model.BoardRoleCommenter,
-		SchemeViewer:    board.MinimumRole == model.BoardRoleViewer,
-	}
-
-	auditRec := a.makeAuditRecord(r, "joinBoard", audit.Fail)
-	defer a.audit.LogRecord(audit.LevelModify, auditRec)
-	auditRec.AddMeta("boardID", boardID)
-	auditRec.AddMeta("addedUserID", userID)
-
-	member, err := a.app.AddMemberToBoard(newBoardMember)
-	if err != nil {
-		a.errorResponse(w, r.URL.Path, http.StatusInternalServerError, "", err)
-		return
-	}
-
-	a.logger.Debug("JoinBoard",
-		mlog.String("boardID", board.ID),
-		mlog.String("addedUserID", userID),
-	)
-
-	data, err := json.Marshal(member)
-	if err != nil {
-		a.errorResponse(w, r.URL.Path, http.StatusInternalServerError, "", err)
-		return
-	}
-
-	// response
-	jsonBytesResponse(w, http.StatusOK, data)
-
-	auditRec.Success()
-}
-
-func (a *API) handleLeaveBoard(w http.ResponseWriter, r *http.Request) {
-	// swagger:operation POST /boards/{boardID}/leave leaveBoard
-	//
-	// Remove your own membership from a board
-	//
-	// ---
-	// produces:
-	// - application/json
-	// parameters:
-	// - name: boardID
-	//   in: path
-	//   description: Board ID
-	//   required: true
-	//   type: string
-	// security:
-	// - BearerAuth: []
-	// responses:
-	//   '200':
-	//     description: success
-	//   '404':
-	//     description: board not found
-	//   '403':
-	//     description: access denied
-	//   default:
-	//     description: internal error
-	//     schema:
-	//       "$ref": "#/definitions/ErrorResponse"
-
-	userID := getUserID(r)
-	if userID == "" {
-		a.errorResponse(w, r.URL.Path, http.StatusBadRequest, "", nil)
-		return
-	}
-
-	boardID := mux.Vars(r)["boardID"]
-
-	if !a.permissions.HasPermissionToBoard(userID, boardID, model.PermissionViewBoard) {
-		a.errorResponse(w, r.URL.Path, http.StatusForbidden, "", nil)
-		return
-	}
-
-	board, err := a.app.GetBoard(boardID)
-	if err != nil {
-		a.errorResponse(w, r.URL.Path, http.StatusInternalServerError, "", err)
-		return
-	}
-	if board == nil {
-		a.errorResponse(w, r.URL.Path, http.StatusNotFound, "", nil)
-		return
-	}
-
-	auditRec := a.makeAuditRecord(r, "leaveBoard", audit.Fail)
-	defer a.audit.LogRecord(audit.LevelModify, auditRec)
-	auditRec.AddMeta("boardID", boardID)
-	auditRec.AddMeta("addedUserID", userID)
-
-	err = a.app.DeleteBoardMember(boardID, userID)
-	if errors.Is(err, app.ErrBoardMemberIsLastAdmin) {
-		a.errorResponse(w, r.URL.Path, http.StatusBadRequest, "", err)
-		return
-	}
-	if err != nil {
-		a.errorResponse(w, r.URL.Path, http.StatusInternalServerError, "", err)
-		return
-	}
-
-	a.logger.Debug("LeaveBoard",
-		mlog.String("boardID", board.ID),
-		mlog.String("addedUserID", userID),
-	)
-
-	jsonStringResponse(w, http.StatusOK, "{}")
-
-	auditRec.Success()
-}
-
-func (a *API) handleUpdateMember(w http.ResponseWriter, r *http.Request) {
-	// swagger:operation PUT /boards/{boardID}/members/{userID} updateMember
-	//
-	// Updates a board member
-	//
-	// ---
-	// produces:
-	// - application/json
-	// parameters:
-	// - name: boardID
-	//   in: path
-	//   description: Board ID
-	//   required: true
-	//   type: string
-	// - name: userID
-	//   in: path
-	//   description: User ID
-	//   required: true
-	//   type: string
-	// - name: Body
-	//   in: body
-	//   description: membership to replace the current one with
-	//   required: true
-	//   schema:
-	//     "$ref": "#/definitions/BoardMember"
-	// security:
-	// - BearerAuth: []
-	// responses:
-	//   '200':
-	//     description: success
-	//     schema:
-	//       $ref: '#/definitions/BoardMember'
-	//   default:
-	//     description: internal error
-	//     schema:
-	//       "$ref": "#/definitions/ErrorResponse"
-
-	boardID := mux.Vars(r)["boardID"]
-	paramsUserID := mux.Vars(r)["userID"]
-	userID := getUserID(r)
-
-	requestBody, err := ioutil.ReadAll(r.Body)
-	if err != nil {
-		a.errorResponse(w, r.URL.Path, http.StatusInternalServerError, "", err)
-		return
-	}
-
-	var reqBoardMember *model.BoardMember
-	if err = json.Unmarshal(requestBody, &reqBoardMember); err != nil {
-		a.errorResponse(w, r.URL.Path, http.StatusBadRequest, "", err)
-		return
-	}
-
-	newBoardMember := &model.BoardMember{
-		UserID:          paramsUserID,
-		BoardID:         boardID,
-		SchemeAdmin:     reqBoardMember.SchemeAdmin,
-		SchemeEditor:    reqBoardMember.SchemeEditor,
-		SchemeCommenter: reqBoardMember.SchemeCommenter,
-		SchemeViewer:    reqBoardMember.SchemeViewer,
-	}
-
-	if !a.permissions.HasPermissionToBoard(userID, boardID, model.PermissionManageBoardRoles) {
-		a.errorResponse(w, r.URL.Path, http.StatusForbidden, "", PermissionError{"access denied to modify board members"})
-		return
-	}
-
-	auditRec := a.makeAuditRecord(r, "patchMember", audit.Fail)
-	defer a.audit.LogRecord(audit.LevelModify, auditRec)
-	auditRec.AddMeta("boardID", boardID)
-	auditRec.AddMeta("patchedUserID", paramsUserID)
-
-	member, err := a.app.UpdateBoardMember(newBoardMember)
-	if errors.Is(err, app.ErrBoardMemberIsLastAdmin) {
-		a.errorResponse(w, r.URL.Path, http.StatusBadRequest, "", err)
-		return
-	}
-	if err != nil {
-		a.errorResponse(w, r.URL.Path, http.StatusInternalServerError, "", err)
-		return
-	}
-
-	a.logger.Debug("PatchMember",
-		mlog.String("boardID", boardID),
-		mlog.String("patchedUserID", paramsUserID),
-	)
-
-	data, err := json.Marshal(member)
-	if err != nil {
-		a.errorResponse(w, r.URL.Path, http.StatusInternalServerError, "", err)
-		return
-	}
-
-	// response
-	jsonBytesResponse(w, http.StatusOK, data)
-
-	auditRec.Success()
-}
-
-func (a *API) handleDeleteMember(w http.ResponseWriter, r *http.Request) {
-	// swagger:operation DELETE /boards/{boardID}/members/{userID} deleteMember
-	//
-	// Deletes a member from a board
-	//
-	// ---
-	// produces:
-	// - application/json
-	// parameters:
-	// - name: boardID
-	//   in: path
-	//   description: Board ID
-	//   required: true
-	//   type: string
-	// - name: userID
-	//   in: path
-	//   description: User ID
-	//   required: true
-	//   type: string
-	// security:
-	// - BearerAuth: []
-	// responses:
-	//   '200':
-	//     description: success
-	//   '404':
-	//     description: board not found
-	//   default:
-	//     description: internal error
-	//     schema:
-	//       "$ref": "#/definitions/ErrorResponse"
-
-	boardID := mux.Vars(r)["boardID"]
-	paramsUserID := mux.Vars(r)["userID"]
-	userID := getUserID(r)
-
-	board, err := a.app.GetBoard(boardID)
-	if err != nil {
-		a.errorResponse(w, r.URL.Path, http.StatusInternalServerError, "", err)
-		return
-	}
-	if board == nil {
-		a.errorResponse(w, r.URL.Path, http.StatusNotFound, "", err)
-		return
-	}
-
-	if !a.permissions.HasPermissionToBoard(userID, boardID, model.PermissionManageBoardRoles) {
-		a.errorResponse(w, r.URL.Path, http.StatusForbidden, "", PermissionError{"access denied to modify board members"})
-		return
-	}
-
-	auditRec := a.makeAuditRecord(r, "deleteMember", audit.Fail)
-	defer a.audit.LogRecord(audit.LevelModify, auditRec)
-	auditRec.AddMeta("boardID", boardID)
-	auditRec.AddMeta("addedUserID", paramsUserID)
-
-	deleteErr := a.app.DeleteBoardMember(boardID, paramsUserID)
-	if errors.Is(deleteErr, app.ErrBoardMemberIsLastAdmin) {
-		a.errorResponse(w, r.URL.Path, http.StatusBadRequest, "", deleteErr)
-		return
-	}
-	if deleteErr != nil {
-		a.errorResponse(w, r.URL.Path, http.StatusInternalServerError, "", deleteErr)
-		return
-	}
-
-	a.logger.Debug("DeleteMember",
-		mlog.String("boardID", boardID),
-		mlog.String("addedUserID", paramsUserID),
-	)
-
-	// response
-	jsonStringResponse(w, http.StatusOK, "{}")
-
 	auditRec.Success()
 }
